requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
langchain==0.1.4
langchain-groq==0.0.1
langchain-community==0.0.16
groq==0.4.2
pinecone-client==3.0.2
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
aiofiles==23.2.1